rateeq.py: potential performance improvements
I've added some code changes in how the evolution matrix is constructed for the rate equations.  Most of these code changes involve breaking out the the construction of the decay portion of the rate equations.  For matrices whose submatrix is diagonal, there is no need to be constantly reconstructing this portion of the matrix.

SPE