common.py: progressBar() updates
Fixed bugs in progressBar():

1. Now only updates once every 0.5 s by default.
2. Now works with standard 80 column wide terminals.

The bug that was preventing it from working in standard terminals was that it was printing more than 80 characters so it was going over one line.  The carriage return only brings it back to the start of the line.  So if one sets the arguments for the progressBar() in such a way that it overfills the terminal line, the bug will reappear.  We might want to consider a more permanent fix.

SPE