Update 04_Fn_to_Fm_1D_MOT.ipynb
Update the example to include the evolution of the test particle.

In order to really narrow in on when the instablility comes in, we need to  vary the recoil velocity, which is not trivial because we also need to vary the detuning.  Figuring out a consistent way to do both might be interesting for this example.

SPE